Improved documentaion
